Fix bag and checkout total

+++ bag/contexts.py
@@ -13,7 +13,8 @@ def bag_contents(request):
 
     for item_id, quantity in bag.items():
         book = get_object_or_404(Book, pk=item_id)
-        total += quantity * float(book.price)
+        total += quantity * Decimal(book.price)
+
         product_count += quantity
         bag_items.append({
             'item_id': item_id,
@@ -22,13 +23,13 @@ def bag_contents(request):
         })
 
     if total < settings.FREE_DELIVERY_THRESHOLD:
-        delivery = int(total) * Decimal(settings.STANDARD_DELIVERY_PERCENTAGE / 100)
+        delivery = total * Decimal(settings.STANDARD_DELIVERY_PERCENTAGE / 100)
         free_delivery_delta = settings.FREE_DELIVERY_THRESHOLD - total
     else:
         delivery = 0
         free_delivery_delta = 0
 
-    grand_total = delivery + int(total)
+    grand_total = delivery + total
 
     context = {
         'bag_items': bag_items,
